Change String to char arrays to save memory
The current implementation seems to require more memory than what the
Arduino Uno has. Changing String objects to char arrays decreased the
use of memory, but the module is still running into execution issues
when the display and the LED light strip are used at the same time. Code
seems to need more refactoring to run in an Uno.