Change button release parameter for game balance
For the hold and wait function in the button module, invert the use of
number 1 and 5 in the parameters of when to release the button to make
it balanced compared to the original game.
Reasoning: In the original game (with a countdown) the number 1 occurs
once every ten secons for the first 4 minutes of game, but with a normal
timer, after 60 seconds the number 1 is constantly in the screen for 1
minute, changing the rareness of finding a 1 to release the button.